15-11-2021

+++ src/appointment/appointment.resolver.ts
@@ -6,17 +6,19 @@ import { UpdateAppointmentInput } from './dto/update-appointment.input';
 import { getApptsDTO } from './dto/get-app.dto';
 import { User } from '../model/user.model';
 import { PrismaService } from '../prisma/prisma.service';
-import { UseGuards, UsePipes } from '@nestjs/common';
+import { UsePipes, NotFoundException, Logger } from '@nestjs/common';
 import { ValidationPipe } from 'src/middleware_logger/validation.pipe';
-import { AuthGaurd } from '../auth/auth.gaurd';
-import { Appointment } from '@prisma/client';
 import MyContext  from '../types/myContext';
+import { LoggerService } from '../logger/logger.service';
+import { log_form } from '../middleware_logger/log_form';
 
 
 @Resolver(() => Appt)
 export class AppointmentResolver {
   constructor(private readonly apptService: AppointmentService,
     private prisma: PrismaService) {}
+    private readonly logger: LoggerService = new Logger(AppointmentResolver.name);
+
 
   @Query(() => Appt, { name: 'get_appointment_by_id' })
   // @UseGuards(new AuthGaurd())
@@ -24,10 +26,20 @@ export class AppointmentResolver {
     return this.apptService.appointment(context.req.user, id);
   }
 
-  // @Query(() => [Appt], { name: 'all_appointments' })
-  // async appointments() {
-  //   return this.apptService.appointments();
-  // }
+  @Query(() => [Appt], { name: 'all_appointments' })
+  async appointments(@Context() context: MyContext) {
+    if(context.req.user === undefined) {
+      this.logger.error(
+        log_form(
+          'all_appointments',
+          `User have not logged in `,
+          new Date().toDateString(),
+        ),
+      );
+      throw new NotFoundException(`You must be logged in to system `);
+    }
+    return this.apptService.appointments(context.req.user);
+  }
 
   @Query(() => Appt, { name: 'appointmentsByUser' })
   @UsePipes(new ValidationPipe(AppointmentResolver))
@@ -35,12 +47,20 @@ export class AppointmentResolver {
     return this.apptService.appointmentsByUser(context.req.user, input);
   }
 
-
-
   
   @Mutation(() => Appt, { name: 'createAppt' })
   @UsePipes(new ValidationPipe(AppointmentResolver))
   async create(@Context() context: MyContext, @Args('input') input: CreateAppointmentInput) {
+    if(context.req.user === undefined) {
+      this.logger.error(
+        log_form(
+          'createAppt',
+          `User have not logged in `,
+          new Date().toDateString(),
+        ),
+      );
+      throw new NotFoundException(`You must be logged in to system `);
+    }
     return this.apptService.createAppt(context.req.user, input);
   }
 
@@ -51,18 +71,37 @@ export class AppointmentResolver {
     @Args('id') id: string,
     @Args('input') args: UpdateAppointmentInput,
   ) {
+    if(context.req.user === undefined) {
+      this.logger.error(
+        log_form(
+          'updateAppt',
+          `User have not logged in `,
+          new Date().toDateString(),
+        ),
+      );
+      throw new NotFoundException(`You must be logged in to system `);
+    }
     return this.apptService.updateAppt(context.req.user ,id, args);
   }
 
   @Mutation(() => Appt, { name: 'deleteAppt' })
   @UsePipes(new ValidationPipe(AppointmentResolver))
   async delete(@Context() context: MyContext,@Args('id') id: string) {
+    if(context.req.user === undefined) {
+      this.logger.error(
+        log_form(
+          'deleteAppt',
+          `User have not logged in `,
+          new Date().toDateString(),
+        ),
+      );
+      throw new NotFoundException(`You must be logged in to system `);
+    }
     return this.apptService.deleteAppt(context.req.user,id);
   }
 
   @ResolveField(()=> User, {nullable: true})
     async user(@Parent() appointments : Appt ){
-       console.log("Dang o day nay");    
         return this.prisma.appointment.findUnique({where: {id: appointments.id}})
     }
 }
